requirements.txt
transformers
torch
torchvision
torchaudio
numpy
scipy
scikit-image
opencv-python
pillow
diffusers
accelerate
matplotlib
tqdm
config
einops
onnxruntime
basicsr
av
fvcore
cloudpickle
omegaconf
pycocotools
